refactor consignments into a separate file
